requirements.txt
strands-agents==1.4.0
strands-agents-tools==0.2.3
strands-agents-mcp==0.2.3
uvx==0.2.3
uvx-stdio==0.2.3
uvx-bedrock37==0.2.3